refactor : dto 패키지 통합
서비스에서 사용하는 DTO와 컨트롤러에서 사용하는 DTO의 내부 필드가 다른 경우보다 같은 것이 합당하기 때문에 통합했다.
