--- src/main/java/com/example/demo/controller/dto/package-info.java
@@ -1,7 +1,0 @@
-/**
- * <p>
- *     Controller 의 RequestMapping 계열의 어노테이션이 붙은 메서드의 매개변수로 사용되는 DTO 클래스가 속한 패키지다.
- *     controller가 아닌 다른 패키지에서 사용하면 안된다.
- * </p>
- */
-package com.example.demo.controller.dto;
