+++ src/main/java/com/example/demo/dto/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.example.demo.controller.dto;
+package com.example.demo.dto;
 
 import lombok.Data;
