+++ src/main/java/com/example/demo/dto/PagedLogicResult.java
@@ -1,4 +1,4 @@
-package com.example.demo.service.dto;
+package com.example.demo.dto;
 
 import java.util.List;
