+++ src/main/java/com/example/demo/dto/StudySearchBase.java
@@ -1,4 +1,4 @@
-package com.example.demo.service.dto;
+package com.example.demo.dto;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
