+++ src/main/java/com/example/demo/dto/StudySearchResult.java
@@ -1,4 +1,4 @@
-package com.example.demo.service.dto;
+package com.example.demo.dto;
 
 public class StudySearchResult { //ToDo 구현 필요
 }
