--- src/main/java/com/example/demo/service/dto/package-info.java
@@ -1,6 +1,0 @@
-/**
- * <p>
- *     service 패키지에 속한 여러 Service 의 메서드의 매개변수 혹은 반환으로 사용되는 클래스가 속한 패키지다.
- * </p>
- */
-package com.example.demo.service.dto;
